arrange import module in train.py move pycolor to config.py
